% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bc.R
\docType{package}
\name{Bc}
\alias{Bc}
\title{Bc.}
\description{
A package performs DNA Barcode analysis. The package can handle all kinds of
barcodes, as long as the barcode withine a single reads have a pattern can be
matched by regular expression. And it can handle barcode with flexible
length, and with or without UMI (unique molecular identifier).
}
\details{
This tool also can be used for preprocssing of amplicon data analysis such as
CRISPR gRNA screening, immue repotoire sequencing and meta genome data.
}
