% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{bc2genBarCode}
\alias{bc2genBarCode}
\title{Generate genBarCode object from Bc BarcodeObj}
\usage{
bc2genBarCode(barcodeObj, BC_backbone = NULL, resDir = getwd())
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{BC_backbone}{A string, the backbone of the barcode}

\item{resDir}{A string, the location to store the data of genBarCode object}
}
\value{
A list, contains genBarCode object
}
\description{
Use the uncorrected barcode to generate the genBarCode data object
}
