% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBarcode.R
\name{bc_cure}
\alias{bc_cure}
\title{Correct the PCR and sequencing mutation in the barcode sequence}
\usage{
bc_cure(x, ...)
}
\description{
Correct the PCR and sequencing mutation in the barcode sequence
}
