% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBarcode.R
\name{bc_cure_depth}
\alias{bc_cure_depth}
\title{Filter barcode by count}
\usage{
bc_cure_depth(barcodeObj, depth = 0, isUpdate = TRUE)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{depth}{Numeric, only a sequence with counts larger than the threshold will be kept.}

\item{isUpdate}{Bool. The function will use cleanBc if it exists, when it's false, the messyBc will be used instead cleanBc to perform the filtering. In the case that no cleanBc is available, messyBc is used.}
}
\value{
A BarcodeObj
}
\description{
The function will perform the filter on the reads count or the UMI counts.
}
\examples{
data(bc_obj)

d1 <- data.frame(
 seq = c(
   "ACTTCGATCGATCGAAAAGATCGATCGATC",
   "AATTCGATCGATCGAAGAGATCGATCGATC",
   "CCTTCGATCGATCGAAGAAGATCGATCGATC",
   "TTTTCGATCGATCGAAAAGATCGATCGATC",
   "AAATCGATCGATCGAAGAGATCGATCGATC",
   "CCCTCGATCGATCGAAGAAGATCGATCGATC",
   "GGGTCGATCGATCGAAAAGATCGATCGATC",
   "GGATCGATCGATCGAAGAGATCGATCGATC",
   "ACTTCGATCGATCGAACAAGATCGATCGATC",
   "GGTTCGATCGATCGACGAGATCGATCGATC",
   "GCGTCCATCGATCGAAGAAGATCGATCGATC"
   ),
 freq = c(
   30, 60, 9, 10, 14, 5, 10, 30, 6, 4 , 6
   )
 )

pattern <- "([ACTG]{3})TCGATCGATCGA([ACTG]+)ATCGATCGATC"
bc_obj <- bc_extract(list(test = d1), pattern, sample_name=c("test"), 
  pattern_type=c(UMI=1, barcode=2))

# Remove barcodes with depth <= 5
(bc_cured <- bc_cure_depth(bc_obj, depth=5))
bc_2matrix(bc_cured)

# Do the clustering, integre the less abundent barcodes to the more abundent one by hammering 
# distance <= 1
bc_cure_cluster(bc_cured, distance = 1)

# Use UMI information to filter the barcode <= 5 UMI-barcode tags
bc_umi_cured <- bc_cure_umi(bc_obj, depth =0, doFish=TRUE, isUniqueUMI=TRUE)
bc_cure_depth(bc_umi_cured, depth = 5)

}
