% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcode.R
\name{bc_extract.character}
\alias{bc_extract.character}
\title{Extract barcode from character vector}
\usage{
\method{bc_extract}{character}(
  file,
  pattern = "",
  sample_name = basename(file),
  meta_data = NULL,
  maxLDist = 0,
  pattern_type = c(barcode = 1),
  costs = list(sub = 1, ins = 99, del = 99),
  ordered = T,
  verbose = T
)
}
\description{
Extract barcode from character vector
}
