% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcode.R
\name{bc_extract.integer}
\alias{bc_extract.integer}
\title{Extract barcode from integer vector}
\usage{
\method{bc_extract}{integer}(
  x,
  pattern = "",
  sample_name = NULL,
  maxLDist = 0,
  pattern_type = c(barcode = 1),
  costs = list(sub = 1, ins = 99, del = 99),
  ordered = T,
  verbose = T
)
}
\description{
Extract barcode from integer vector
}
