% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcode.R
\name{bc_extract.list}
\alias{bc_extract.list}
\title{Extract barcode from list}
\usage{
\method{bc_extract}{list}(
  x,
  pattern = "",
  sample_name = NULL,
  meta_data = NULL,
  maxLDist = 0,
  pattern_type = c(barcode = 1),
  costs = list(sub = 1, ins = 99, del = 99),
  ordered = T,
  verbose = F
)
}
\description{
Extract barcode from list
}
