% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bc_filterSeq.character}
\alias{bc_filterSeq.character}
\title{Filter the reads by the QC threshold for character vector}
\usage{
\method{bc_filterSeq}{character}(
  x,
  min_average_quality = 30,
  min_read_length = 0,
  N_threshold = 0,
  sample_name = basename(x)
)
}
\description{
Filter the reads by the QC threshold for character vector
}
