% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bc_filterSeq.data.frame}
\alias{bc_filterSeq.data.frame}
\title{Filter the reads by the QC threshold for data.frame}
\usage{
\method{bc_filterSeq}{data.frame}(x, min_read_length = 0, N_threshold = 0)
}
\description{
Filter the reads by the QC threshold for data.frame
}
