% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bc_filterSeq.integer}
\alias{bc_filterSeq.integer}
\title{Filter the reads by the QC threshold for integer vector}
\usage{
\method{bc_filterSeq}{integer}(x, min_read_length = 0, N_threshold = 0)
}
\description{
Filter the reads by the QC threshold for integer vector
}
