% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bc_filterSeq.list}
\alias{bc_filterSeq.list}
\title{Filter the reads by the QC threshold for list}
\usage{
\method{bc_filterSeq}{list}(
  x,
  min_average_quality = 30,
  min_read_length = 0,
  N_threshold = 0,
  sample_name = names(x)
)
}
\description{
Filter the reads by the QC threshold for list
}
