% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bc_filterSeq.ShortReadQ}
\alias{bc_filterSeq.ShortReadQ}
\title{Filter the reads by the QC threshold}
\usage{
\method{bc_filterSeq}{ShortReadQ}(x, min_average_quality = 30, min_read_length = 0, N_threshold = 0)
}
\description{
Filter the reads by the QC threshold
}
