% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_meta}
\alias{bc_meta}
\alias{bc_meta<-}
\title{Metadata in Barcode}
\usage{
bc_meta(barcodeObj)

bc_meta(barcodeObj, key = NULL) <- value
}
\arguments{
\item{barcodeObj}{A BarcodeObj.}

\item{key}{A string, the meta data record name.}

\item{value}{A string vector with the length of sample number in the
BarcodeObj, or a data.frame with the length of the sample number, the new
meta data record value.  If a data.frame is given, the rownames of the
data.frame should be the sample names.}
}
\value{
A data.frame
}
\description{
Function to get or set metadata in BarcodeObj.
}
\examples{
data(bc_obj)

# get the metadata data.frame
bc_meta(bc_obj)

# assign value to metadata by $ operation
bc_meta(bc_obj)$phenotype <- c("l", "b")

# assign value to metasta by key argument
bc_meta(bc_obj, key = "sample_type") <- c("l", "b")

# show the updated metadata
bc_meta(bc_obj)

# assign a new data.frame as metadata
metadata <- data.frame(
    sample_name <- c("test1", "test2"),
    phenotype <- c("l", "b")
    )
rownames(metadata) = bc_names(bc_obj)
bc_meta(bc_obj) <- metadata
###
}
