% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_names}
\alias{bc_names}
\alias{bc_names<-}
\title{Access & update sample names in BarcodeObj}
\usage{
bc_names(barcodeObj)

bc_names(barcodeObj) <- value
}
\arguments{
\item{barcodeObj}{A \code{BarcodeObj} object.}

\item{value}{A character vector setting the new sample names, with the length of the samples number in
\code{BarcodeObj} object.}
}
\value{
A character vector
}
\description{
Get or update sample names in BarcodeObj object.
}
\examples{
data(bc_obj)

bc_names(bc_obj)
bc_names(bc_obj) <- c("new1", "new2")
}
