% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBarcode.R
\name{bc_plotNetwork}
\alias{bc_plotNetwork}
\title{Draw Network graph to show the similarity of barcode}
\usage{
bc_plotNetwork(x, ...)
}
\description{
Draw Network graph to show the similarity of barcode
}
