% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_runQC.character}
\alias{bc_runQC.character}
\title{Perform quality control for character vector}
\usage{
\method{bc_runQC}{character}(file, sample_name = basename(file))
}
\description{
Perform quality control for character vector
}
