% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_runQC.DNAStringSet}
\alias{bc_runQC.DNAStringSet}
\title{Perform quality control for DNAStringSet}
\usage{
\method{bc_runQC}{DNAStringSet}(x, plot = F)
}
\description{
Perform quality control for DNAStringSet
}
