% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_runQC.ShortReadQ}
\alias{bc_runQC.ShortReadQ}
\title{Perform quality control for ShortReadQ}
\usage{
\method{bc_runQC}{ShortReadQ}(x, n = 50, plot = F)
}
\description{
Perform quality control for ShortReadQ
}
