% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_seqQC}
\alias{bc_seqQC}
\alias{bc_seqQC.ShortReadQ}
\alias{bc_seqQC.DNAStringSet}
\alias{bc_seqQC.data.frame}
\alias{bc_seqQC.integer}
\alias{bc_seqQC.character}
\alias{bc_seqQC.list}
\alias{plot.barcodeQc}
\alias{plot.barcodeQcSet}
\title{Evaluates sequences quality}
\usage{
bc_seqQC(...)

\method{bc_seqQC}{ShortReadQ}(x, ...)

\method{bc_seqQC}{DNAStringSet}(x, ...)

\method{bc_seqQC}{data.frame}(x, ...)

\method{bc_seqQC}{integer}(x, ...)

\method{bc_seqQC}{character}(file, sample_name = basename(file), ...)

\method{bc_seqQC}{list}(x, sample_name = names(x), ...)

\method{plot}{barcodeQc}(x, ...)

\method{plot}{barcodeQcSet}(x, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{x}{A single or list of Fastq file, ShortReadQ object, DNAStringSet
object, data.frame or named integer vector.}

\item{file}{A character vector, specifying the input Fastq file.}

\item{sample_name}{A character vector with the length of sample number, used
to set the sample name.}
}
\value{
A barcodeQc or a barcodeQcSet class. 
The barcodeQc is a list with four elements, 
\itemize{
  \item \code{top}: a \code{data.frame} with top 50 most frequency sequence, 
  \item \code{distribution}: a \code{data.frame} with the distribution of read depth. It contains 
  \code{nOccurrences} (depth), and \code{nReads} (unique sequence) columns.
  \item \code{base_quality_per_cycle}: \code{data.frame} with base-pair
  location (NGS sequencing cycle) by row, and the base-pair quality summary by column,
  including Mean, P5 (5% quantile), P25 (25% quantile), Median, P75 (75%
  quantile) and P95 (95% quantile).
  \item \code{base_freq_per_cycle}: \code{data.frame} with three columns: 1. \code{Cycle}, the
  sequence base-pair location (NGS sequencing cycle); 2. \code{Base}, DNA base;
  \code{Count}: reads count.
  \item{summary}: a numeric vector with following elements:
  \code{total_read}, \code{median_read_length},
  \code{p5_read_length}, \code{p95_read_length}.
}
The barcodeQcSet is a list of barcodeQc.
}
\description{
\code{bc_seqQC} evaluates sequences quality. See the return value for detail.
}
\examples{
library(ShortRead)
# fastq file
fq_file <- system.file("extdata", "simple.fq", package="CellBarcode")
bc_seqQC(fq_file)

# ShortReadQ
sr <- readFastq(fq_file[1])
bc_seqQC(sr)

# DNAStringSet
ds <- sr@sread
bc_seqQC(ds)

# List of DNAStringSet
l <- list(sample1 = ds, sample2 = ds)
plot(bc_seqQC(l))

# List of ShortRead
l_sr <- list(sample1 = sr, sample2 = sr)
plot(bc_seqQC(l_sr))

###
}
