% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_subset}
\alias{bc_subset}
\alias{[.BarcodeObj}
\alias{+.BarcodeObj}
\alias{-.BarcodeObj}
\alias{*.BarcodeObj}
\title{Subset operation of BarcodeObj object}
\usage{
bc_subset(
  barcodeObj,
  sample = NULL,
  barcode = NULL,
  black_list = NULL,
  is_sample_quoted_exp = FALSE
)

\method{[}{BarcodeObj}(barcodeObj, barcode = NULL, sample = NULL, ...)

\method{+}{BarcodeObj}(barcodeObj_x, barcodeObj_y)

\method{-}{BarcodeObj}(barcodeObj, black_list)

\method{*}{BarcodeObj}(barcodeObj, white_list)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{sample}{A vector or an expression, select the sample subset from BarcodeObj}

\item{barcode}{A vector of integer or string, select the barcode subset from BarcodeObj}

\item{black_list}{A character vector. Remove the listed barcodes.}

\item{is_sample_quoted_exp}{A bool value. If the sample parameter is quoted expression or not}

\item{...}{Additional arguments}

\item{barcodeObj_x}{A BarcodeObj}

\item{barcodeObj_y}{A BarcodeObj}

\item{white_list}{A character vector. Only keep the listed barcodes.}
}
\value{
A BarcodeObj
}
\description{
Subset operation of BarcodeObj object
}
\examples{
data(bc_obj)

bc_obj

# Select barcodes
bc_subset(bc_obj, barcode = c("AACCTT", "AACCTT"))
bc_obj[c("AGAG", "AAAG"), ]

# Select by meta data
bc_meta(bc_obj)$phenotype <- c("l", "b")
bc_meta(bc_obj)
bc_subset(bc_obj, phenotype == "l")

bc_obj[, phenotype == "l"]

# Select samples by sample name
bc_obj[, "test1"]
bc_obj[, c("test1", "test2")]
bc_subset(bc_obj, sample = "sample1_rep1", barcode = c("AACCTT", "AACCTT"))

# Apply black list
bc_subset(bc_obj, sample = c("sample1_rep1", "sample1_rep2"), barcode = c("AACCTT"))

# Join two samples with different barcode sets
bc_obj["AGAG", "test1"] + bc_obj["AAAG", "test1"]

# Join two samples with overlap barcodes
bc_obj_join <- bc_obj["AGAG", "test1"] + bc_obj["AGAG", "test1"]
bc_obj_join
# The same barcode will merged after applying bc_cure_depth()
bc_cure_depth(bc_obj_join)

# Remove barcodes
bc_obj
bc_obj - "AAAG"

# Select barcode in white list
bc_obj
bc_obj * "AAAG"
###
}
