% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_summary_qc}
\alias{bc_summary_qc}
\title{Summary barcodeQcSet}
\usage{
bc_summary_qc(x)
}
\arguments{
\item{x}{a barcodeQcSet object.}
}
\value{
A \code{data.frame} with 5 columns: \code{sample_name},
\code{total_read}, \code{median_read_length}, \code{p5_read_length} and
\code{p95_read_length.}
}
\description{
Summary the "total read count" and "read length" of each samples within
a \code{BarcodeQcSet} object, and output a \code{data.frame} with sample by row and different
metrics by column.
}
\examples{

fq_file <- dir(
    system.file("extdata", "mef_test_data", package = "CellBarcode"),
    full=TRUE)

summary(bc_seqQC(fq_file))
###

}
