% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellBarcode.R
\docType{_PACKAGE}
\name{CellBarcode}
\alias{CellBarcode}
\title{DNA Barcode Analysis toolkit}
\description{
The package CellBarcode performs Cellular DNA Barcode analysis. It can handle all kinds of DNA barcodes, as long as the barcode is within a single sequencing read and has a pattern that can be matched by a regular expression. \code{CellBarcode} can handle barcodes with flexible lengths, with or without UMI (unique molecular identifier). This tool also can be used for pre-processing some amplicon data such as CRISPR gRNA screening, immune repertoire sequencing, and metagenome data.
}
