% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descripBarcode.R
\name{countBc.BarcodeObj}
\alias{countBc.BarcodeObj}
\title{Readout the barcode count}
\usage{
\method{countBc}{BarcodeObj}(barcodeObj)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}
}
\value{
A list, the item is a data.frame, with column names of `barcode_seq` and `count`, the list name is sample names
}
\description{
Readout the barcode count
}
