% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cureBc.BarcodeObj}
\alias{cureBc.BarcodeObj}
\title{Correct the PCR and sequencing mutation in the barcode sequence}
\usage{
\method{cureBc}{BarcodeObj}(
  barcodeObj,
  depth_threshold = 0,
  barcode_count_threshold = 1000,
  hammer_dist_threshold = 1,
  with_umi = F
)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{depth_threshold}{A integer, only a barcode with sequencing depth larger than that will be processed}

\item{with_umi}{A bool value, True when UMI is used}

\item{hamming_dist_threshold}{A integer, the hamming distance threshold of two distinct barcode}
}
\value{
A BarcodeObj
}
\description{
Correct the PCR and sequencing mutation in the barcode sequence
}
