% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBarcode.R
\name{cureBc}
\alias{cureBc}
\title{Correct the PCR and sequencing mutation in the barcode sequence}
\usage{
cureBc(x, ...)
}
\description{
Correct the PCR and sequencing mutation in the barcode sequence
}
