% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{-.BarcodeObj}
\alias{-.BarcodeObj}
\title{Remove the barcode in black list}
\usage{
\method{-}{BarcodeObj}(x, y)
}
\arguments{
\item{x}{A BarcodeObj}

\item{y}{A string or vector, the barcode black list to revmoe}
}
\value{
A BarcodeObj
}
\description{
Remove the barcode in black list
}
