% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{filterFq.BarcodeObj}
\alias{filterFq.BarcodeObj}
\title{Filter the reads by the QC threshold}
\usage{
\method{filterFq}{BarcodeObj}(barcodeObj, min_quality = 20, min_mean_quality = 30, min_length = 30)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{min_quality}{A number. The minimum reads base quality threshold}

\item{min_mean_quality}{A number. The minimum average reads base quality threshold}

\item{min_length}{A number. The minimum length of the reads}
}
\value{
A BarcodeObj
}
\description{
Filter the reads by the QC threshold
}
