% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterFq.ShortReadQ}
\alias{filterFq.ShortReadQ}
\title{Filter the reads by the QC threshold}
\usage{
\method{filterFq}{ShortReadQ}(x, min_average_quality = 30, min_read_length = 0)
}
\description{
Filter the reads by the QC threshold
}
