% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{format.BarcodeObj}
\alias{format.BarcodeObj}
\title{Format BarcodeObj}
\usage{
\method{format}{BarcodeObj}(x, ...)
}
\arguments{
\item{x}{A BarcodeObj}

\item{...}{Additional arguments}
}
\description{
Format the summary of BarcodeObj for pretty print.
}
\examples{
data(bc_obj)

# format BarcodeObj for pretty print
format(bc_obj)

###
}
\seealso{
\code{\link[Bc::print.BarcodeObj]{Bc::print.BarcodeObj}}
}
