% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BarcodeObj.R
\name{format,BarcodeObj-method}
\alias{format,BarcodeObj-method}
\title{Formats BarcodeObj object}
\usage{
\S4method{format}{BarcodeObj}(x)
}
\arguments{
\item{x}{A BarcodeObj object}
}
\value{
Formated summary text.
}
\description{
Format the summary of BarcodeObj object for pretty print.
}
\examples{
data(bc_obj)

# format BarcodeObj for pretty print
format(bc_obj)

###
}
