% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parse_10x_sam}
\alias{parse_10x_sam}
\title{Parse 10X SAM file}
\usage{
parse_10x_sam(in_file_path, regex_str, cell_barcode_tag, umi_tag)
}
\arguments{
\item{in_file_path}{A string, define the un-mapped sequences}

\item{regex_str}{A string, define the regular expression to match the barcode sequence.}

\item{cell_barcode_tag}{A string, define the tag of 10X cell barcode field in sam file.}

\item{umi_tag}{A string, define the tag of UMI field in the sam file.}
}
\value{
A list containing two data frames:
  - `barcode_df`: with columns `cell_barcode`, `umi`, `barcode_seq`, and `count`.
  - `raw_reads_df`: with columns `cell_barcode` and `count`.
}
\description{
Parse 10X SAM file
}
