% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{plot.barcodeQc}
\alias{plot.barcodeQc}
\title{Visualize sequencing quality for single sample}
\usage{
\method{plot}{barcodeQc}(x, opt = "all")
}
\description{
Visualize sequencing quality for single sample
}
