% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{plot.barcodeQcSet}
\alias{plot.barcodeQcSet}
\title{Visualize sequencing quality for several samples}
\usage{
\method{plot}{barcodeQcSet}(x)
}
\description{
Visualize sequencing quality for several samples
}
