% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{plotNetwork.BarcodeObj}
\alias{plotNetwork.BarcodeObj}
\title{Draw Network graph to show the similarity of barcode}
\usage{
\method{plotNetwork}{BarcodeObj}(
  barcodeObj,
  sample1 = NULL,
  sample2 = NULL,
  topN = 100,
  type = "raw",
  UMI_depth_threshold_n = 5,
  ...
)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{sample1}{A string, sample name}

\item{sample2}{A string or NULL, the second sample while comparing the barcode between two samples}

\item{topN}{A integer, only the barcodes with top N depth or UMI-barcode count are used}

\item{type}{A integer, "raw" or "clean" correspond to the uncorrected or corrected barcde data}

\item{UMI_depth_threshold_n}{A integer, only the barcode with read depth or UMI-barcode count greater than the value will be used}
}
\description{
Draw Network graph to show the similarity of barcode
}
