% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{+.BarcodeObj}
\alias{+.BarcodeObj}
\title{Combined the BarcodeObj}
\usage{
\method{+}{BarcodeObj}(x, y)
}
\arguments{
\item{x}{A BarcodeObj}

\item{y}{A BarcodeObj}
}
\value{
A BarcodeObj
}
\description{
Combined the BarcodeObj
}
