% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{print.BarcodeObj}
\alias{print.BarcodeObj}
\title{Print out the summary of BarcodeObj}
\usage{
\method{print}{BarcodeObj}(barcodeObj)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}
}
\value{
The general information about the BarcodeObj
}
\description{
Print out the summary of BarcodeObj
}
