% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{qcSummary.BarcodeObj}
\alias{qcSummary.BarcodeObj}
\title{Visualize the Quality information of Fastq file}
\usage{
\method{qcSummary}{BarcodeObj}(barcodeObj)
}
\description{
Visualize the Quality information of Fastq file
}
