% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{readFq}
\alias{readFq}
\title{Read the fastq file}
\usage{
readFq(fastqFiles, sampleNames)
}
\arguments{
\item{fastqFiles}{A vector or string. The fastq file location}

\item{sampleNames}{A vector or string, which should be the same length to the `fastqFiles`. The sample name corespond to the input fastq files}
}
\value{
A BarcodeObj
}
\description{
Read the fastq file
}
