% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{runQC}
\alias{runQC}
\title{get the QC information of the raw sequence
fastq file name, ShortReadQ Obj, DNAStringSet Obj, data.frame, vector}
\usage{
runQC(...)
}
\description{
get the QC information of the raw sequence
fastq file name, ShortReadQ Obj, DNAStringSet Obj, data.frame, vector
}
