% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{samplenames.BarcodeObj}
\alias{samplenames.BarcodeObj}
\title{Get the sample names from BarcodeObj}
\usage{
\method{samplenames}{BarcodeObj}(barcodeObj)
}
\arguments{
\item{barcodeObj}{BarcodeObj}
}
\value{
A string vector
}
\description{
Get the sample names from BarcodeObj
}
