% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seq_correct}
\alias{seq_correct}
\title{Sequence clustering using Levenshtein or Hamming distance}
\usage{
seq_correct(
  seq,
  count,
  count_threshold,
  dist_threshold,
  depth_fold_threshold,
  dist_method,
  cost
)
}
\arguments{
\item{seq}{A string vector of sequences.}

\item{count}{An integer vector corresponding to sequence counts.}

\item{count_threshold}{A Integer Threshold to consider a sequence as true barcode.}

\item{dist_threshold}{A Integer Distance threshold for clustering.}

\item{depth_fold_threshold}{Fold change threshold for merging.}

\item{dist_method}{Distance method: 2 for Levenshtein, otherwise Hamming.}

\item{cost}{A vector of three integers for Levenshtein distance costs, the order is insert, delete, replace.}
}
\value{
A list with two data frames:
- `seq_freq_tab`: Corrected sequence frequency table.
- `link_tab`: Clustering process record.
}
\description{
Sequence clustering using Levenshtein or Hamming distance
}
