% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BarcodeObj.R, R/method-seqQc.R
\name{show,BarcodeObj-method}
\alias{show,BarcodeObj-method}
\alias{show,BarcodeQc-method}
\alias{show,BarcodeQcSet-method}
\title{Show BarcodeObj object}
\usage{
\S4method{show}{BarcodeObj}(object)

\S4method{show}{BarcodeQc}(object)

\S4method{show}{BarcodeQcSet}(object)
}
\arguments{
\item{object}{A BarcodeQcSet object}
}
\value{
Formated summary text.

Formated summary text.

Formated summary text.
}
\description{
Show the summary of BarcodeObj object for pretty print.

Show the summary of BarcodeQc object for pretty print.

Show the summary of BarcodeQcSet object for pretty print.
}
\examples{
data(bc_obj)

# show BarcodeObj for pretty print
bc_obj

###
}
