% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{stairPlot.BarcodeObj}
\alias{stairPlot.BarcodeObj}
\title{Stair plot used to find a optimized cutoff point}
\usage{
\method{stairPlot}{BarcodeObj}(barcodeObj, sample = 1)
}
\arguments{
\item{barcodeObj}{A BarcodeObj with extracted barcode.}
}
\description{
Stair plot used to find a optimized cutoff point
}
