% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{[.BarcodeObj}
\alias{[.BarcodeObj}
\title{Subset operation of BarcodeObj object}
\usage{
\method{[}{BarcodeObj}(barcodeObj, x = NULL, y = NULL)
}
\arguments{
\item{x}{A vector or string. Select the barcode subset from BarcodeObj}

\item{y}{A vector or string. Select the sample subset from BarcodeObj}
}
\value{
A BarcodeObj
}
\description{
Subset operation of BarcodeObj object
}
\examples{
# Not run
#bc_obj[c("AACCTT", "AACCTT"), ]
#bc_obj[, "sample1_rep1"]
#bc_obj[, c("sample1_rep1", "sample1_rep2")]
}
