% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{subset.BarcodeObj}
\alias{subset.BarcodeObj}
\title{Subset operation of BarcodeObj object}
\usage{
\method{subset}{BarcodeObj}(
  barcodeObj,
  sample = NULL,
  barcode = NULL,
  black_list = NULL,
  sample_is_expression = F
)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{sample}{A vector of integer or string, select the barcode subset from BarcodeObj}

\item{barcode}{A vector or an expression, select the sample subset from BarcodeObj}

\item{black_list}{A vector or string. Remove the the barcode in the list.}
}
\value{
A BarcodeObj
}
\description{
Subset operation of BarcodeObj object
}
\examples{
# Not run
# subset(bc_obj, barcode = c("AACCTT", "AACCTT"))
# subset(bc_obj, sample = "sample1_rep1", barcode = c("AACCTT", "AACCTT"))
# subset(bc_obj, sample = c("sample1_rep1", "sample1_rep2"), barcode = c("AACCTT", "AACCTT"))
}
