% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{*.BarcodeObj}
\alias{*.BarcodeObj}
\title{Select the barcode by white list}
\usage{
\method{*}{BarcodeObj}(x, y)
}
\arguments{
\item{x}{A BarcodeObj}

\item{y}{A string or vector, the barcode white list}
}
\value{
A BarcodeObj
}
\description{
Select the barcode by white list
}
