% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bc.R
\docType{package}
\name{Bc}
\alias{Bc}
\title{DNA Barcode Analysis toolkit}
\description{
This package performs DNA Barcode analysis. The package can handle all kinds
of DNA barcodes, as long as the barcode within a single reads and has a
pattern that can be matched by a regular expression. This package can handle
barcode with flexible length, with or without UMI (unique molecular
identifier). This tool also can be used for preprocessing of amplicon data
such as CRISPR gRNA screening, immune repertoire sequencing and meta
genome data.
}
