% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_2df}
\alias{bc_2df}
\alias{bc_2dt}
\alias{bc_2matrix}
\title{Transforms BarcodeObj object into other data type}
\usage{
bc_2df(barcodeObj)

bc_2dt(barcodeObj)

bc_2matrix(barcodeObj)
}
\arguments{
\item{barcodeObj}{A BarcodeObj.}
}
\value{
A data.frame, with two columns named `barcode_seq` and `count`.

A data.table, with two columns named `barcode_seq` and `count`.

A data matrix with sample by columns and barcode by rows.
}
\description{
Transforms BarcodeObj object into data.frame, data.table or matrix.
}
\examples{
data(bc_obj)

bc_obj <- bc_cure_depth(bc_obj)

# get data.frame
bc_2df(bc_obj)

# get data.table
bc_2dt(bc_obj)

# get matrix
bc_2matrix(bc_obj)

###
}
