% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBarcode.R
\name{bc_auto_cutoff}
\alias{bc_auto_cutoff}
\title{Finds barcode count cutoff point}
\usage{
bc_auto_cutoff(barcodeObj, useCleanBc = TRUE)
}
\arguments{
\item{barcodeObj}{A \code{BarcodeObj} object.}

\item{useCleanBc}{A logical value, if \code{TRUE}, the \code{cleanBc} element in the
\code{BarcodeObj} object will be used, otherwise the \code{messyBc} element will be used.}
}
\value{
a numeric \code{vector} of the cutoff point.
}
\description{
Finds the cutoff point for the barcode count filtering based on the barcode
count distribution.
}
\details{

}
\examples{

data(bc_obj)

bc_auto_cutoff(bc_obj)

}
