% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_barcodes}
\alias{bc_barcodes}
\title{Gets barcode sequences}
\usage{
bc_barcodes(barcodeObj, unlist = TRUE)
}
\arguments{
\item{barcodeObj}{A BarcodeObj object.}

\item{unlist}{A logical value. If TRUE, the function returns a vector
of unique barcode list from all samples; otherwise a list will be
returned, with each element containing the barcodes of each sample.}
}
\value{
A character vector or a list.
}
\description{
\code{bc_barcodes} used to get the barcode sequences in BarcodeObj object. The input 
BarcodesObj object should be preprocessed by \code{bc_cure_*} functions.
}
\examples{
data(bc_obj)

# get unique barcode vector
bc_barcodes(bc_obj)

# get barcode for each sample
bc_barcodes(bc_obj, unlist = FALSE)

###
}
