% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_barcodes}
\alias{bc_barcodes}
\title{Get barcode sequences}
\usage{
bc_barcodes(barcodeObj, unlist = TRUE)
}
\arguments{
\item{barcodeObj}{A BarcodeObj.}

\item{unlist}{A bool value, if TRUE, then the function will return a vector of barcodes; else a list will be returned, each element is corresponding to a sample.}
}
\value{
A character vector or a list
}
\description{
Get barcode sequences
}
\examples{
data(bc_obj)

bc_barcodes(bc_obj)

bc_barcodes(bc_obj, unlist = FALSE)
}
