% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_barcodes}
\alias{bc_barcodes}
\title{Get barcode sequences}
\usage{
bc_barcodes(barcodeObj, unlist = TRUE)
}
\arguments{
\item{barcodeObj}{A BarcodeObj.}

\item{unlist}{A bool value, if TRUE, then the function will return a vector
of unique barcodes in all samples; otherwise a list will be returned, each
element is corresponding to a sample.}
}
\value{
A character vector or a list.
}
\description{
This function retrieves the barcode sequences from cleanBc element in
BarcodeObj.
}
\examples{
data(bc_obj)

# get unique barcode vector
bc_barcodes(bc_obj)

# get barcode for each sample
bc_barcodes(bc_obj, unlist = FALSE)

###
}
