% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBarcode.R
\name{bc_cure}
\alias{bc_cure}
\title{Clean barcode sequences}
\usage{
bc_cure(
  barcodeObj,
  depth_threshold = 0,
  with_umi = FALSE,
  umi_depth = 2,
  doFish = FALSE,
  isUniqueUMI = FALSE,
  hammer_dist_threshold = 0,
  barcode_count_threshold = 10000
)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{depth_threshold}{A Numeric, only a sequence with reads depth larger than that will be processed. If a UMI is used, the depth_threshold is the reads threshold for the UMI-sequence amplicon.}

\item{with_umi}{A bool value, True when UMI is applied.}

\item{umi_depth}{A Numeric, the minimum UMI-sequence amplicon counts needed for consider a true sequence.}

\item{doFish}{A bool value, the UMI-sequence amplicon unsatisfied the depth_threshold are counted when when the sequence is in the final sequence list which satisfied the umi_depth threshold.}

\item{isUniqueUMI}{A bool value, True if UMI is considered absolute unique, if so only the dominant sequence is chosen for each UMI.}

\item{hammer_dist_threshold}{A integer, the less abundent sequence are compared to the most abundent ones and the hammering distance is calculated, the less abundent seuqneces with hammering distance less than the value will be removed.}

\item{barcode_count_threshold}{A integer, appliable when hammer_dist_threshold > 0, the max sequences number should be, when the number of sequence are satisfied all the rest less abundent sequences are removed.}
}
\value{
A BarcodeObj
}
\description{
Clean barcode sequences
}
\examples{
data(bc_obj)

d1 <- data.frame(
 seq = c(
   "ACTTCGATCGATCGAAAAGATCGATCGATC",
   "AATTCGATCGATCGAAGAGATCGATCGATC",
   "CCTTCGATCGATCGAAGAAGATCGATCGATC",
   "TTTTCGATCGATCGAAAAGATCGATCGATC",
   "AAATCGATCGATCGAAGAGATCGATCGATC",
   "CCCTCGATCGATCGAAGAAGATCGATCGATC",
   "GGGTCGATCGATCGAAAAGATCGATCGATC",
   "GGATCGATCGATCGAAGAGATCGATCGATC",
   "ACTTCGATCGATCGAACAAGATCGATCGATC",
   "GGTTCGATCGATCGACGAGATCGATCGATC",
   "GCGTCCATCGATCGAAGAAGATCGATCGATC"
   ),
 freq = c(
   30, 60, 9, 10, 14, 5, 10, 30, 6, 4 , 6
   )
 )

pattern <- "([ACTG]{3})TCGATCGATCGA([ACTG]+)ATCGATCGATC"
bc_obj <- bc_extract(list(test = d1), pattern, sample_name=c("test"), 
  pattern_type=c(UMI=1, barcode=2))

# Remove barcodes with depth <= 5
(bc_cured <- bc_cure(bc_obj, depth_threshold=5))
bc_2matrix(bc_cured)

# Do the clustering, integre the less abundent barcodes to the more abundent one by hammering 
# distance <= 1
bc_cure(bc_obj, depth_threshold=5, hammer_dist_threshold = 1)

# Use UMI information to filter the barcode <= 5 UMI-barcode tags
bc_cure(bc_obj, depth_threshold=0, doFish=FALSE, with_umi=TRUE, umi_depth=5, isUniqueUMI=TRUE)

# Use UMI information to filter the barcode <= 5 UMI-barcode tags
# Meanwhile, take into acount the UMI-barcode tags that do not meet the reads depth threshold 
# but contains the true barcode
bc_cure(bc_obj, depth_threshold=0, doFish=TRUE, with_umi=TRUE, umi_depth=5, isUniqueUMI=FALSE)
}
