% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBarcode.R
\name{bc_cure_cluster}
\alias{bc_cure_cluster}
\title{Merge barcodes by editing distance}
\usage{
bc_cure_cluster(
  barcodeObj,
  distance = 1,
  dist_method = "hammer",
  merge_method = "greedy",
  barcode_n = 10000
)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{distance}{Integer, the editing distance threshold if two sequence is similar enough to be merged}

\item{dist_method}{A String, currently it only be "hammer". The hammering distance is used to evaluate the similarity between two barcodes.}

\item{merge_method}{A String, currently only "greedy" is available. Comparing the least abundent barcode to the most abundent ones, merge the least abundent to the most abundent ones.}

\item{barcode_n}{Integer, the max sequences number should be. When the most abundent barcodes are satisfied this number the merging finished, and all the rest sub-abundent sequences are removed.}
}
\description{
This function do the clustering to merge the barcodes with similar sequence. This function only appliable to the BarcodeObj with cleanBc element.
}
\examples{
data(bc_obj)

d1 <- data.frame(
 seq = c(
   "ACTTCGATCGATCGAAAAGATCGATCGATC",
   "AATTCGATCGATCGAAGAGATCGATCGATC",
   "CCTTCGATCGATCGAAGAAGATCGATCGATC",
   "TTTTCGATCGATCGAAAAGATCGATCGATC",
   "AAATCGATCGATCGAAGAGATCGATCGATC",
   "CCCTCGATCGATCGAAGAAGATCGATCGATC",
   "GGGTCGATCGATCGAAAAGATCGATCGATC",
   "GGATCGATCGATCGAAGAGATCGATCGATC",
   "ACTTCGATCGATCGAACAAGATCGATCGATC",
   "GGTTCGATCGATCGACGAGATCGATCGATC",
   "GCGTCCATCGATCGAAGAAGATCGATCGATC"
   ),
 freq = c(
   30, 60, 9, 10, 14, 5, 10, 30, 6, 4 , 6
   )
 )

pattern <- "([ACTG]{3})TCGATCGATCGA([ACTG]+)ATCGATCGATC"
bc_obj <- bc_extract(list(test = d1), pattern, sample_name=c("test"), 
  pattern_type=c(UMI=1, barcode=2))

# Remove barcodes with depth <= 5
(bc_cured <- bc_cure_depth(bc_obj, depth=5))
bc_2matrix(bc_cured)

# Do the clustering, integre the less abundent barcodes to the more abundent one by hammering 
# distance <= 1
bc_cure_cluster(bc_cured, distance = 1)

# Use UMI information to filter the barcode <= 5 UMI-barcode tags
bc_umi_cured <- bc_cure_umi(bc_obj, depth =0, doFish=TRUE, isUniqueUMI=TRUE)
bc_cure_depth(bc_umi_cured, depth = 5)

}
