% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBarcode.R
\name{bc_diversity}
\alias{bc_diversity}
\title{Evaluate barcode diversity}
\usage{
bc_diversity(barcodeObj, plot = TRUE)
}
\arguments{
\item{barcodeObj}{A BarcodeObj object.}

\item{plot}{A logical value, if TRUE, draw the Lorenz curve and barcode
distribution curve.}
}
\value{
A data.frame with following columns:
\itemize{
  \item total_reads: total read number.
  \item uniq_barcode: how many barcodes in the dataset, shows the richness of
  the barcode.
  \item shannon_index: Shannon's diversity index or Shannon–Wiener index.
  \item equitability_index: Shannon's equitability.
  \item bit_index: Shannon bit information.
}
}
\description{
bc_diversity evaluates sequence diversity metrics in the cleanBc
element of BarcodeObj object. Lorenz curve and
barcode frequency distribution figure can be generated by this function.
}
\details{
Followings are the metrics used for evaluating the barcode diversity:

\emph{Richness}: The unique barcodes number \eqn{R} evaluates the richness of
the barcodes.

\emph{Shannon index}: Shannon diversity index is weighted geometric
average of the proportion of barcodes.
\deqn{ H' = - \sum_{i=1}^{R}p_ilnp_i }

\emph{Equitability index}: Shannon equitability \eqn{E_H} characterize the
evenness of thee barcodes, it is a value between 0 and 1, with 1 being
complete evenness.
\deqn{ E_H = H' / H'_{max} = H / ln(R) }

\emph{Bit}:
Shannon entropy \eqn{H}, with a units of bit, 
\deqn{ H = - \sum_{i=1}^{R}p_ilog_2p_i }
}
\examples{
data(bc_obj)

# filter barcode by depth
bc_obj <- bc_cure_depth(bc_obj)

# evaluate barcode diversity
bc_diversity(bc_obj)
}
