% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcode.R
\name{bc_extract}
\alias{bc_extract}
\title{Extract barcode from reads}
\usage{
bc_extract(x, ...)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{pattern}{A string. The regular expression to match the barcode which capture pattern}

\item{maxLDist}{A integer. The mismatch threshold for barcode matching}

\item{pattern_type}{A vector. It defines the barcode (and UMI) capture pattern}

\item{costs}{A vector. Define the weight for each mismatch events}
}
\value{
A barcodeObj
}
\description{
Extract barcode from reads
}
