% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_meta}
\alias{bc_meta}
\alias{bc_meta<-}
\title{Access metadata in BarcodeObj}
\usage{
bc_meta(barcodeObj)

bc_meta(barcodeObj, key = NULL) <- value
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{key}{A string, the meta data record name}

\item{value}{A string vector, the meta data record value}
}
\value{
A BarcodeObj
}
\description{
Access metadata in BarcodeObj
}
\examples{
data(bc_obj)

# get the metadata data.frame
bc_meta(bc_obj)

# assign value to metadata by $ operation
bc_meta(bc_obj)$phenotype <- c("l", "b")

# assign value to metasta by key argument
bc_meta(bc_obj, key = "sample_type") <- c("l", "b")

# show the updated metadata
bc_meta(bc_obj)

# assign a new data.frame as metadata
metadata <- data.frame(
  sample_name <- c("test1", "test2"),
  phenotype <- c("l", "b")
  )
rownames(metadata) = bc_names(bc_obj)
bc_meta(bc_obj) <- metadata
###
}
