% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_meta}
\alias{bc_meta}
\alias{bc_meta<-}
\title{Accesses and sets metadata in BarcodeObj object}
\usage{
bc_meta(barcodeObj)

bc_meta(barcodeObj, key = NULL) <- value
}
\arguments{
\item{barcodeObj}{A BarcodeObj object.}

\item{key}{A string, identifying the metadata record name to be modified.}

\item{value}{A string vector or a data.frame. When the "value" is a vector, it
should have the length of sample number in the BarcodeObj object and the
"key" variable is necessary. Otherwise, when the "value" is data.frame, the
row name of the data.frame should be the sample name, and each column as a
metadata record.}
}
\value{
A data.frame
}
\description{
Accesses and sets metadata in BarcodeObj object
}
\examples{
data(bc_obj)

# get the metadata data.frame
bc_meta(bc_obj)

# assign value to metadata by $ operation
bc_meta(bc_obj)$phenotype <- c("l", "b")

# assign value to metasta by key argument
bc_meta(bc_obj, key = "sample_type") <- c("l", "b")

# show the updated metadata
bc_meta(bc_obj)

# assign a new data.frame as metadata
metadata <- data.frame(
    sample_name <- c("test1", "test2"),
    phenotype <- c("l", "b")
    )
rownames(metadata) = bc_names(bc_obj)
bc_meta(bc_obj) <- metadata
###
}
