% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_names}
\alias{bc_names}
\alias{bc_names<-}
\title{Access sample names in BarcodeObj}
\usage{
bc_names(barcodeObj)

bc_names(barcodeObj) <- value
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{value}{A string vector, the new sample names}
}
\value{
A character vector
}
\description{
Access sample names in BarcodeObj
}
\examples{
data(bc_obj)

bc_names(bc_obj)
bc_names(bc_obj) <- c("new1", "new2")
}
