% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_names}
\alias{bc_names}
\alias{bc_names<-}
\title{Sample names in BarcodeObj}
\usage{
bc_names(barcodeObj)

bc_names(barcodeObj) <- value
}
\arguments{
\item{barcodeObj}{A BarcodeObj.}

\item{value}{A string vector with the length of the samle number. The new
sample names.}
}
\value{
A character vector
}
\description{
Function to get or set sample name.
}
\examples{
data(bc_obj)

bc_names(bc_obj)
bc_names(bc_obj) <- c("new1", "new2")
}
