% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_seqQC}
\alias{bc_seqQC}
\alias{bc_seqQC.ShortReadQ}
\alias{bc_seqQC.DNAStringSet}
\alias{bc_seqQC.data.frame}
\alias{bc_seqQC.integer}
\alias{bc_seqQC.character}
\alias{bc_seqQC.list}
\alias{plot.barcodeQc}
\alias{plot.barcodeQcSet}
\title{Performs quality control}
\usage{
bc_seqQC(...)

\method{bc_seqQC}{ShortReadQ}(x, ...)

\method{bc_seqQC}{DNAStringSet}(x, ...)

\method{bc_seqQC}{data.frame}(x, ...)

\method{bc_seqQC}{integer}(x, ...)

\method{bc_seqQC}{character}(file, sample_name = basename(file), ...)

\method{bc_seqQC}{list}(x, ...)

\method{plot}{barcodeQc}(x, ...)

\method{plot}{barcodeQcSet}(x, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{x}{A single or list of Fastq file, ShortReadQ object, DNAStringSet
object, data.frame or named integer vector.}

\item{file}{A string. Fastq file name}

\item{sample_name}{A string vector with the length of sample number, rename
the samples.}
}
\value{
A barcodeQc or barcodeQcSet class. 
The barcodeQc is a list with two elements, 
\itemize{
  \item base_quality_per_cycle: data.frame with sequence base-pair
  location by row, and the base-pair sequencing quality summary by column.
  \item base_freq_per_cycle: data.frame with three columns, \emph{Cycle}: the
  sequence base-pair location (NGS sequencing cycle); \emph{Base}: DNA base;
  \emph{Count}: reads count.
}
The barcodeQcSet is a list of barcodeQc.
}
\description{
bc_seqQC evaluates the sequence quality.
}
\examples{
library(ShortRead)
# fastq file
fq_file <- system.file("extdata", "simple.fq", package="Bc")
bc_seqQC(fq_file)

# ShortReadQ
sr <- readFastq(fq_file[1])
bc_seqQC(sr)

# DNAStringSet
ds <- sr@sread
bc_seqQC(ds)

# List of DNAStringSet
l <- list(sample1 = ds, sample2 = ds)
plot(bc_seqQC(l))

# List of ShortRead
l_sr <- list(sample1 = sr, sample2 = sr)
plot(bc_seqQC(l_sr))

###
}
