% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{bc_seqQC}
\alias{bc_seqQC}
\alias{bc_seqQC.ShortReadQ}
\alias{bc_seqQC.DNAStringSet}
\alias{bc_seqQC.data.frame}
\alias{bc_seqQC.integer}
\alias{bc_seqQC.character}
\alias{bc_seqQC.list}
\alias{plot.barcodeQc}
\alias{plot.barcodeQcSet}
\title{Perform quality control}
\usage{
bc_seqQC(...)

\method{bc_seqQC}{ShortReadQ}(x)

\method{bc_seqQC}{DNAStringSet}(x)

\method{bc_seqQC}{data.frame}(x)

\method{bc_seqQC}{integer}(x)

\method{bc_seqQC}{character}(file, sample_name = basename(file))

\method{bc_seqQC}{list}(x)

\method{plot}{barcodeQc}(x, opt = "all")

\method{plot}{barcodeQcSet}(x)
}
\arguments{
\item{x}{A string. fastq file name, ShortReadQ Obj, DNAStringSet Obj, data.frame or integer vector}
}
\value{
A barcodeQc or barcodeQcSet class. 
The barcodeQc is list with two elements, 
  base_quality_per_cycle: data.frame with row of sequence basepair location, the columns are the basepair sequencing quality summary
  base_freq_per_cycle: data.frame with three columns, Cycle: the sequence basepair location (NGS sequencing cycle); Base: DNA base; Count: reads count.
The barcodeQcSet is list of barcodeQc
}
\description{
Perform quality control
}
\examples{
fq_file = system.file("extdata", "simple.fq", package="Bc")
sr = readFastq(fq_file)
ds = sr@sread
l = list(sample1 = ds, sample2 = ds)
l_sr = list(sample1 = sr, sample2 = sr)

bc_seqQC(fq_file)
bc_seqQC(sr)
bc_seqQC(ds)
plot(bc_seqQC(l))
plot(bc_seqQC(l_sr))

}
