% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{bc_subset}
\alias{bc_subset}
\alias{[.BarcodeObj}
\alias{+.BarcodeObj}
\alias{-.BarcodeObj}
\alias{*.BarcodeObj}
\title{Manage barcodes or samples in a BarcodeObj}
\usage{
bc_subset(
  barcodeObj,
  sample = NULL,
  barcode = NULL,
  black_list = NULL,
  is_sample_quoted_exp = FALSE
)

\method{[}{BarcodeObj}(barcodeObj, barcode = NULL, sample = NULL, ...)

\method{+}{BarcodeObj}(barcodeObj_x, barcodeObj_y)

\method{-}{BarcodeObj}(barcodeObj, black_list)

\method{*}{BarcodeObj}(barcodeObj, white_list)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{sample}{A vector or an expression, specifying the samples in the
subsets. And in the expression, the metadata can be used as variable.}

\item{barcode}{A vector of integer or string, indicating the selected
barcode.}

\item{black_list}{A character vector. Remove the barcodes in the black list.}

\item{is_sample_quoted_exp}{A bool value. If TRUE, indicating the
expression in \emph{sample} argument is quoted.}

\item{...}{Additional arguments}

\item{barcodeObj_x}{A BarcodeObj}

\item{barcodeObj_y}{A BarcodeObj}

\item{white_list}{A character vector. Only barcodes in the white list kept.}
}
\value{
A BarcodeObj
}
\description{
Get a subsets of or join BarcodesObj.
}
\details{
`bc_subset` and `[]` used for selecting samples barcodes.

`+` used for combining two BarcodesObj.

`-` used for removing barcodes in a black_list.

`*` used for selecting barcodes in a white_list.
}
\examples{
data(bc_obj)

bc_obj

# Select barcodes
bc_subset(bc_obj, barcode = c("AACCTT", "AACCTT"))
bc_obj[c("AGAG", "AAAG"), ]

# Select samples by meta data
bc_meta(bc_obj)$phenotype <- c("l", "b")
bc_meta(bc_obj)
bc_subset(bc_obj, phenotype == "l")

bc_obj[, phenotype == "l"]

# Select samples by sample name
bc_obj[, "test1"]
bc_obj[, c("test1", "test2")]
bc_subset(bc_obj, sample = "sample1_rep1", barcode = c("AACCTT", "AACCTT"))

# Apply barcodes black list
bc_subset(
bc_obj,
    sample = c("sample1_rep1", "sample1_rep2"),
    barcode = c("AACCTT"))

# Join two samples with different barcode sets
bc_obj["AGAG", "test1"] + bc_obj["AAAG", "test1"]

# Join two samples with overlap barcodes
bc_obj_join <- bc_obj["AGAG", "test1"] + bc_obj["AGAG", "test1"]
bc_obj_join
# The same barcode will merged after applying bc_cure_depth()
bc_cure_depth(bc_obj_join)

# Remove barcodes
bc_obj
bc_obj - "AAAG"

# Select barcodes in white list
bc_obj
bc_obj * "AAAG"
###
}
