% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellBarcode.R
\docType{package}
\name{CellBarcode}
\alias{CellBarcode}
\title{DNA Barcode Analysis toolkit}
\description{
This package performs DNA Barcode (genetic lineage tracing) analysis. The
package can handle all kinds of DNA barcodes, as long as the barcode within a
single sequencing read and has a pattern which can be matched by a regular
expression. This package can handle barcode with flexible length, with or
without UMI (unique molecular identifier). This tool also can be used for
pre-processing of some amplicon data such as CRISPR gRNA screening, immune
repertoire sequencing and meta genome data.
}
