% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{format.BarcodeObj}
\alias{format.BarcodeObj}
\title{Formats BarcodeObj object}
\usage{
\method{format}{BarcodeObj}(x, ...)
}
\arguments{
\item{x}{A BarcodeObj}

\item{...}{Additional arguments}
}
\value{
Formated BarcodeObj summary.
}
\description{
Formats the summary of BarcodeObj object for pretty print.
}
\examples{
data(bc_obj)

# format BarcodeObj for pretty print
format(bc_obj)

###
}
\seealso{
\code{\link[CellBarcode]{print.BarcodeObj}}
}
