% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{format.BarcodeObj}
\alias{format.BarcodeObj}
\title{Format the summary of BarcodeObj}
\usage{
\method{format}{BarcodeObj}(x, ...)
}
\arguments{
\item{x}{A BarcodeObj}

\item{...}{Additional arguments}
}
\description{
Format the summary of BarcodeObj
}
\examples{
data(bc_obj)

format(bc_obj)
print(bc_obj)
}
