% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{print.BarcodeObj}
\alias{print.BarcodeObj}
\title{Print the summary of BarcodeObj}
\usage{
\method{print}{BarcodeObj}(barcodeObj)
}
\description{
Print the summary of BarcodeObj
}
\examples{
data(bc_obj)

format(bc_obj)
print(bc_obj)
}
