% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{print.BarcodeObj}
\alias{print.BarcodeObj}
\title{Print the summary of BarcodeObj}
\usage{
\method{print}{BarcodeObj}(x, ...)
}
\arguments{
\item{x}{A BarcodeObj}

\item{...}{Additional arguments}
}
\description{
Print the summary of BarcodeObj
}
\examples{
data(bc_obj)

# print BarcodeObj
print(bc_obj)
###
}
\seealso{
[Bc::format.BarcodeObj] format BarcodeObj for pretty print
}
