% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_correct}
\alias{seq_correct}
\title{Sequence clustering}
\usage{
seq_correct(seq, count, count_threshold, hamm_dist_threshold)
}
\arguments{
\item{seq}{A string vector}

\item{count}{An integer vector with the same order and length of UMI}

\item{count_threshold}{An integer the maximum barcodes number}

\item{hamm_dist_threshold}{An integer the hamming distance threshold}
}
\description{
This function will merge the UMIs by using the 
hamming distance. If two UMIs have hamming distance
no more than 1, only the UMI with more reads
will be kept.
}
\details{
This function will return the corrected UMI list.
}
