% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_correct}
\alias{seq_correct}
\title{Sequence clustering}
\usage{
seq_correct(seq, count, count_threshold, hammer_dist_threshold)
}
\arguments{
\item{count}{An integer vector with the same order and 
length of UMI}

\item{umi}{An String vector with UMI}
}
\description{
This function will merge the UMIs by using the 
hammer distance. If two UMIs have hammer distance
no more than 1, only the UMI with more reads
will be kept.
}
\details{
This function will return the corrected UMI list.
}
