% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeObj.R
\name{[.BarcodeObj}
\alias{[.BarcodeObj}
\title{Subset operation of BarcodeObj object}
\usage{
\method{[}{BarcodeObj}(barcodeObj, x = NULL, y = NULL)
}
\arguments{
\item{barcodeObj}{A BarcodeObj}

\item{x}{A vector of integer or string, select the barcode subset from BarcodeObj}

\item{y}{A vector or an expression, select the sample subset from BarcodeObj}
}
\value{
A BarcodeObj
}
\description{
Subset operation of BarcodeObj object
}
\examples{
data(bc_obj)
bc_obj[c("AGAG", "AAAG"), ]
bc_obj[, "test1"]
bc_obj[, sample_name == "test1"]
bc_obj[, c("test1", "test2")]

}
